refactor NavBar
