+++ src/pages/404.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import SEO from '../components/seo/Seo';
 import LayoutWrapper from '../components/layouts/LayoutWrapper';
 import MainPageDarkLayout from '../components/layouts/MainPageDarkLayout';
-import NavBarMainPage from '../components/nav-bar/NavBarMainPage';
+import NavBarMain from '../components/nav-bar/NavBarMain';
 import NavBarHomeButton from '../components/nav-bar/NavBarHomeButton';
 import NavBarContactButton from '../components/nav-bar/NavBarContactButton';
 
@@ -11,8 +11,8 @@ function NotFoundPage() {
     <LayoutWrapper>
       <MainPageDarkLayout
         navBar={
-          <NavBarMainPage
-            className="sticky"
+          <NavBarMain
+            classNameProp="sticky"
             desktopRightContent={
               <div className="flex justify-end items-center">
                 <NavBarHomeButton className="mr-2" />
