+++ src/pages/contact.jsx
@@ -4,7 +4,7 @@ import MainPageDarkLayout from '../components/layouts/MainPageDarkLayout';
 import BusinessCardContactPage from '../components/contact/business-card/BusinessCardContactPage';
 import ContactPageForm from '../components/contact/contact-page-form/ContactPageForm';
 import SEO from '../components/seo/Seo';
-import NavBarMainPage from '../components/nav-bar/NavBarMainPage';
+import NavBarMain from '../components/nav-bar/NavBarMain';
 import NavBarHomeButton from '../components/nav-bar/NavBarHomeButton';
 import { MobileAndTabletQuery, DesktopQuery } from '../util/helpers';
 import LayoutWrapper from '../components/layouts/LayoutWrapper';
@@ -14,10 +14,9 @@ function ContactPage() {
     <LayoutWrapper>
       <MainPageDarkLayout>
         <SEO title="contact.seo.title" description="contact.seo.description" />
-        <NavBarMainPage
+        <NavBarMain
           showAlways
-          showNavBar
-          className="sticky"
+          classNameProp="sticky"
           desktopRightContent={
             <div className="flex justify-end items-center">
               <NavBarHomeButton />
