+++ src/pages/imprint.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import LayoutWrapper from '../components/layouts/LayoutWrapper';
 import MainPageDarkLayout from '../components/layouts/MainPageDarkLayout';
-import NavBarMainPage from '../components/nav-bar/NavBarMainPage';
+import NavBarMain from '../components/nav-bar/NavBarMain';
 import '../css/text-page-styles.scss';
 import NavBarContactButton from '../components/nav-bar/NavBarContactButton';
 import SEO from '../components/seo/Seo';
@@ -10,10 +10,10 @@ import SEO from '../components/seo/Seo';
 const ImprintPage = () => (
   <LayoutWrapper>
     <MainPageDarkLayout>
-      <NavBarMainPage
+      <NavBarMain
         showNavBar
         showAlways
-        className="sticky"
+        classNameProp="sticky"
         desktopRightContent={
           <div className="flex justify-end items-center">
             <NavBarContactButton />
