+++ src/pages/softwarelaboratory.jsx
@@ -5,7 +5,7 @@ import SectionHeading from '../components/index/section-heading/SectionHeading';
 import { DesktopQuery, MobileAndTabletQuery } from '../util/helpers';
 import GradientButton from '../components/buttons/gradient-border-buttons/GradientButton';
 import MainPageDarkLayout from '../components/layouts/MainPageDarkLayout';
-import NavBarMainPage from '../components/nav-bar/NavBarMainPage';
+import NavBarMain from '../components/nav-bar/NavBarMain';
 import NavBarContactButton from '../components/nav-bar/NavBarContactButton';
 import SoftwareLaboratorySectionContent from '../components/index/page-sections/SoftwareLaboratorySectionContent';
 import LayoutWrapper from '../components/layouts/LayoutWrapper';
@@ -16,10 +16,10 @@ function SoftwareLaboratoryPage() {
   return (
     <LayoutWrapper>
       <MainPageDarkLayout>
-        <NavBarMainPage
+        <NavBarMain
           showNavBar
           showAlways
-          className="z-20 sticky w-full navbar-background-blur animate top-0 mb-4"
+          classNameProp="z-20 sticky w-full navbar-background-blur animate top-0 mb-4"
           desktopRightContent={
             <div className="flex justify-end items-center">
               <NavBarContactButton />
